# app/Models/Song.php
<?php

namespace App\Models;

use App\Models\Album;
use Illuminate\Database\Eloquent\Model;

class Song extends Model
{
    public function albums()
    {
        return $this->belongsToMany(Album::class)->withPivot('track_number');
    }
}