# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music Catalog API",
        "description": "API для работы с каталогом исполнителей, альбомов и песен.",
        "contact": {
            "email": "admin@musiccatalog.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/songs": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Получить список всех песен",
                "description": "Возвращает список всех песен в базе данных",
                "operationId": "0388bb28c35eec49b8b492d169ce3575",
                "responses": {
                    "200": {
                        "description": "Успешный ответ с данными песен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Song"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            },
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Создать новую песню",
                "description": "Создает новую песню в базе данных",
                "operationId": "2f1970b82d91cc715578d972189f19cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Song"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Создана новая песня",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Song"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные данные"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        },
        "/api/songs/{id}": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Получить информацию о песне",
                "description": "Возвращает информацию о песне по ее ID",
                "operationId": "4325a96361d1d3820cc0580eaae9197c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID песни",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ с данными песни",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Song"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Песня не найдена"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            },
            "put": {
                "tags": [
                    "Songs"
                ],
                "summary": "Обновить информацию о песне",
                "description": "Обновляет информацию о песне по ее ID",
                "operationId": "452de7f8f11a121cd11e9430565632c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID песни",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Song"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Информация о песне обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Song"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Песня не найдена"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Songs"
                ],
                "summary": "Удалить песню",
                "description": "Удаляет песню по ее ID",
                "operationId": "829c1a52ae5013ce3ff4b9829a31485f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID песни",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Песня удалена"
                    },
                    "404": {
                        "description": "Песня не найдена"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        },
        "/api/songs/{id}/add-to-album": {
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Добавить песню в альбом",
                "description": "Добавляет песню в альбом по ID",
                "operationId": "5645e43aef0d6a587bff7b411c6be977",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID песни",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "album_id": {
                                        "description": "ID альбома",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "track_number": {
                                        "description": "Номер трека в альбоме",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Песня добавлена в альбом",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Song added to album successfully."
                                        },
                                        "album": {
                                            "$ref": "#/components/schemas/Album"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Альбом или песня не найдены"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Album": {
                "required": [
                    "title",
                    "artist_id",
                    "release_year"
                ],
                "properties": {
                    "id": {
                        "description": "ID альбома",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Название альбома",
                        "type": "string",
                        "example": "Abbey Road"
                    },
                    "artist_id": {
                        "description": "ID исполнителя",
                        "type": "integer",
                        "example": 1
                    },
                    "release_year": {
                        "description": "Год выпуска альбома",
                        "type": "integer",
                        "example": 1969
                    },
                    "created_at": {
                        "description": "Дата создания записи",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-01T14:30:00Z"
                    },
                    "updated_at": {
                        "description": "Дата последнего обновления записи",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-01T14:30:00Z"
                    }
                },
                "type": "object"
            },
...
            "Song": {
                "required": [
                    "title",
                    "artist_id",
                    "album_id",
                    "release_date"
                ],
                "properties": {
                    "id": {
                        "description": "ID песни",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Название песни",
                        "type": "string",
                        "example": "Hey Jude"
                    },
                    "artist_id": {
                        "description": "ID исполнителя",
                        "type": "integer",
                        "example": 1
                    },
                    "album_id": {
                        "description": "ID альбома",
                        "type": "integer",
                        "example": 1
                    },
                    "release_date": {
                        "description": "Дата выпуска",
                        "type": "string",
                        "format": "date",
                        "example": "1968-08-26"
                    },
                    "created_at": {
                        "description": "Дата создания записи",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-01T14:30:00Z"
                    },
                    "updated_at": {
                        "description": "Дата последнего обновления записи",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-01T14:30:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}